+++ src/main/use_case/navigation/maplocation/MapLocation.java
@@ -15,13 +15,13 @@ public interface MapLocation {
      * Returns the x-coordinate of the location.
      * @return the x-coordinate of the location
      */
-    double getX();
+    int getX();
 
     /**
      * Returns the y-coordinate of the location.
      * @return the y-coordinate of the location
      */
-    double getY();
+    int getY();
 
     /**
      * Returns the floor ID of the location.
